completed all tests and created signature for throws exception test
