+++ src/integrationTest/java/uk/gov/hmcts/darts/audio/service/AudioTransformationServiceTest.java
@@ -30,8 +30,6 @@ class AudioTransformationServiceTest {
     @Autowired
     private AudioTransformationService audioTransformationService;
 
-    private Path filePath;
-
     private Integer requestId;
 
     @BeforeEach
@@ -58,53 +56,5 @@ class AudioTransformationServiceTest {
     void processAudioRequest() {
         MediaRequestEntity processingMediaRequestEntity = audioTransformationService.processAudioRequest(requestId);
         assertEquals(PROCESSING, processingMediaRequestEntity.getStatus());
-    }
-
-// Extract the following Tests into another Tests Class
-    @Test
-    @DisplayName("Test-1: Check if file is created in temporary folder")
-    void saveBlobDataToTempWorkspaceTest_1() {
-        String data = "this is a binary data file";
-        BinaryData mediaFile = BinaryData.fromString(data);
-        String fileName = "caseAudioFile.pdf";
-        //extract the above three lines in @BeforeEach method
-
-        filePath = audioTransformationService.saveBlobDataToTempWorkspace(mediaFile, fileName);
-
-        assertTrue(Files.exists(filePath));
-    }
-
-    @Test
-    @DisplayName("Test-2: Check if file is empty")
-    void saveBlobDataToTempWorkspaceTest_2() throws IOException {
-        String data = "this is a binary data file";
-        BinaryData mediaFile = BinaryData.fromString(data);
-        String fileName = "caseAudioFile.pdf";
-        //extract the above three lines in @BeforeEach method
-
-        filePath = audioTransformationService.saveBlobDataToTempWorkspace(mediaFile, fileName);
-
-        assertNotEquals(0L, Files.size(filePath));
-    }
-
-    @Test
-    @DisplayName("Test-3: Check if the saved file is equal to the original BinaryData file")
-    void saveBlobDataToTempWorkspaceTest_3() throws IOException {
-        String data = "this is a binary data file";
-        BinaryData mediaFile = BinaryData.fromString(data);
-        String fileName = "caseAudioFile.pdf";
-        //extract the above three lines in @BeforeEach method
-
-        filePath = audioTransformationService.saveBlobDataToTempWorkspace(mediaFile, fileName);
-
-        assertArrayEquals(mediaFile.toBytes(), Files.readAllBytes(filePath));
-    }
-
-    @AfterEach
-    void deleteFile() throws IOException, InterruptedException {
-        if (filePath != null) {
-            Thread.sleep(1000);     //remove this line before pull request
-            Files.delete(filePath);
-        }
     }
 }
