+++ src/main/java/uk/gov/hmcts/darts/audio/service/AudioTransformationService.java
@@ -3,12 +3,13 @@ package uk.gov.hmcts.darts.audio.service;
 import com.azure.core.util.BinaryData;
 import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
 
+import java.io.IOException;
 import java.nio.file.Path;
 
 public interface AudioTransformationService {
 
     MediaRequestEntity processAudioRequest(Integer requestId);
 
-    Path saveBlobDataToTempWorkspace(BinaryData mediaFile, String fileName);
+    Path saveBlobDataToTempWorkspace(BinaryData mediaFile, String fileName) throws IOException;
 
 }
